#1077 * Fixed bug with Tangent command.xml generation
